requirements.txt
google-api-python-client==1.12.1
oauth2client==3.0.0
google-cloud-bigquery==1.27.2
google-cloud-firestore==1.9.0
grpcio==1.32.0